Don't require that users "speak HTTP"
Remove the requirement for users to be familiar with HTTP status codes
to make sense of the dialog. No room in UI for the full sentence "Is the
page missing?" but this will suffice. Since commit 119ef4115a, the
extension handles more situations than just 404 as well – so the
message is not correct as-is anymore.